create a quantum circuit with three qubits
Beginning a hunt for a polynomial time algorithm for quantum graph isomorphism. Start by putting a qubit on each edge of a graph. Then we have a map from (S^3)^N -> R_{\ge 0}^N, where N = (n choose 2) is the number of edges and n is the number of vertices. S^3 is the Bloch sphere, the space of a qubit with two complex components whose sum of magnitudes squared is equal to one. Quotienting by the permutation group \Sigma_n which rearranges vertices (and induces an edge permutation) gives the moduli space of unlabeled networks, both quantum and classical.